basic styling and added photo to homepage
