+++ client/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" href="./assets/TableTalk.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>TableTalk</title>
+    <title>TableTalk- Where Foodies Become Friends</title>
   </head>
   <body>
     <div id="root"></div>
