+++ client/pages/Homepage/Homepage.css
@@ -1,0 +1,31 @@
+body {
+    background-color: whitesmoke;
+    color: black;
+    margin: 0;
+    padding: 0;
+    font-family: Arial, sans-serif;
+  }
+h1 {
+    font-size: xx-large;
+}
+
+p {
+    font-size: larger;
+}
+  .homeCard {
+    background-color: transparent;
+    color: black;
+    font-size: 100%;
+  }
+  .signUp {
+    background-color: #c24646;
+  }
+
+  .howBtn {
+    background-color: #ffa6d7;
+  }
+  @media only screen and (max-width: 600px) {
+    main {
+      padding: 10px;
+    }
+}
