+++ client/pages/index.js
@@ -1,6 +1,7 @@
 
 
 export { default as AuthPage } from "./AuthPage"
-export { default as HomePage } from "./HomePage"
+export { default as HomePage } from "./Homepage/HomePage"
 export { default as Logout } from "./Logout"
 export { default as PrivatePage } from "./ProfilePage"
+
